requirements-backend.txt
# Backend - FastAPI Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.1

# APIs externas
openai==1.3.7
requests==2.31.0
aiohttp==3.9.1

# Autenticación y seguridad
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6

# Utilidades
python-dotenv==1.0.0
pillow==10.1.0
numpy==1.25.2
pandas==2.1.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Desarrollo
black==23.11.0
flake8==6.1.0
mypy==1.7.1